% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_partial_time_compat_lubridate.R
\name{as.interval,partial_time-method}
\alias{as.interval,partial_time-method}
\title{Wrapper for lubridate as.interval}
\usage{
\S4method{as.interval}{partial_time}(x, start, ...)
}
\arguments{
\item{x}{a duration, difftime, period, or numeric object that describes the length of the interval}

\item{start}{a POSIXt or Date object that describes when the interval begins}

\item{...}{additional arguments to pass to as.interval}
}
\value{
A \code{partial_timespan} vector
}
\description{
Wrapper for lubridate as.interval
}
