% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_partial_time_coersion.R
\name{as.parttime}
\alias{as.parttime}
\title{Coerce an object to a parttime object}
\usage{
as.parttime(x)
}
\arguments{
\item{x}{an object for coersion}
}
\description{
Coerce an object to a parttime object
}
