% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_partial_time_c.R
\name{c.partial_time}
\alias{c.partial_time}
\title{Concatenate partial_time objects}
\usage{
\method{c}{partial_time}(...)
}
\arguments{
\item{...}{\code{partial_time} objects to concatenate}
}
\value{
A \code{partial_time} object.
}
\description{
Concatenate partial_time objects
}
\examples{
c(as.parttime("2015-04"), as.parttime("2016-08-01"))

}
