% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_partial_time_dim.R
\name{dim.partial_time}
\alias{dim.partial_time}
\title{parttime vector dimensions}
\usage{
\method{dim}{partial_time}(x)
}
\arguments{
\item{x}{A \code{partial_time} object}
}
\value{
An \code{integer} vector of dimensions (length) of a \code{partial_time} vector
}
\description{
parttime vector dimensions
}
