% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.i}
\alias{.i}
\title{shorthand for converting dimnames to indices, for easier column subtraction}
\usage{
.i(x, dim, ...)
}
\arguments{
\item{x}{an object with dimnames<- defined}

\item{dim}{the dimension to index}

\item{...}{names selected by which}
}
\description{
shorthand for converting dimnames to indices, for easier column subtraction
}
\keyword{internal}
