% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_timespan_compat_stats.R
\name{end}
\alias{end}
\title{end S3 generic}
\usage{
end(x, ...)
}
\arguments{
\item{x}{An object to retrieve the end from}

\item{...}{Additional arguments passed to methods}
}
\description{
A generic method to retrieve the end of an object
}
