% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract}
\alias{extract}
\title{Similar to matrix and array `[` behavior, but allows for providing a numeric
vector of dimensions to drop}
\usage{
extract(x, ..., drop = TRUE, envir = parent.frame())
}
\description{
Similar to matrix and array `[` behavior, but allows for providing a numeric
vector of dimensions to drop
}
