% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_partial_time_format.R
\name{format.partial_time}
\alias{format.partial_time}
\title{Format a parttime object}
\usage{
\method{format}{partial_time}(x, ..., quote = TRUE)
}
\arguments{
\item{x}{A \code{partial_time} object}

\item{...}{Additional arguments passed to \link{format_field_matrix}}

\item{quote}{A \code{logical} indicating whether to add quotation marks around
formatted objects}
}
\value{
A \code{character} vector representation of a \code{partial_time} vector
}
\description{
Format a parttime object
}
