% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_partial_time_compat_pillar.R
\name{format.pillar_shaft_partial_time}
\alias{format.pillar_shaft_partial_time}
\title{parttime pillar formatting}
\usage{
\method{format}{pillar_shaft_partial_time}(x, width, ...)
}
\arguments{
\item{x}{A character vector with formatting,
can use ANYI styles e.g provided by the \pkg{cli} package.}

\item{width}{A maximum display width of the each element in the resulting
vector of strings}

\item{...}{Additional arguments unused}
}
\value{
A \code{character} representation of a \code{partial_time} vector
}
\description{
parttime pillar formatting
}
