% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_partial_time_format.R
\name{format_field_matrix}
\alias{format_field_matrix}
\title{Format individual components of a parttime matrix}
\usage{
format_field_matrix(
  x,
  verbose = getOption("parttime.print_verbose", FALSE),
  tz
)
}
\arguments{
\item{x}{a parttime \code{matrix} to format}

\item{verbose}{a \code{logical} value indicating whether to include full
timestamps, even if partially missing.}

\item{tz}{a \code{logical} value indicating whether to include timezone
information. By default, will only display time zones if they are not
missing and not equal to the assumed timezone option.}
}
\description{
Format individual components of a parttime matrix
}
\keyword{internal}
