% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_partial.R
\name{has_partial_time}
\alias{has_partial_time}
\title{Test whether a partial_time object's time components are incomplete}
\usage{
has_partial_time(x)
}
\arguments{
\item{x}{a partial_time object to test for incompleteness}
}
\value{
A logical vector indicating whether each element of a \code{partial_time}
has any missing time fields.
}
\description{
Test whether a partial_time object's time components are incomplete
}
\examples{
has_partial_time(as.parttime(c("2015-04-13", "2015-04-13T10:30:15")))

}
