% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_time.R
\name{impute_date_max}
\alias{impute_date_max}
\title{Impute with the maximum possible date}
\usage{
impute_date_max(x, ..., res = "day")
}
\arguments{
\item{x}{timestamp object to impute}

\item{...}{additional individual named fields to impute. Can be one of
"year", "month", "day", "hour", "min", "sec", "secfrac", "tzhour", "tzmin".}
}
\description{
Impute with the maximum possible date
}
\seealso{
Other parttime_impute: 
\code{\link{impute_date_mid}()},
\code{\link{impute_date_min}()},
\code{\link{impute_time_max}()},
\code{\link{impute_time_mid}()},
\code{\link{impute_time_min}()},
\code{\link{impute_time}()}
}
\concept{parttime_impute}
