% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_time.R
\name{impute_time_mid}
\alias{impute_time_mid}
\title{Impute with middle timestamp values}
\usage{
impute_time_mid(x, tz = "GMT", ...)
}
\arguments{
\item{x}{timestamp object to impute}

\item{tz}{a timezone to use, defaults to +1400}

\item{...}{additional individual named fields to impute. Can be one of
"year", "month", "day", "hour", "min", "sec", "secfrac", "tzhour", "tzmin".}
}
\description{
Impute with middle timestamp values
}
\seealso{
Other impute_time: 
\code{\link{impute_time_max}()},
\code{\link{impute_time_min}()},
\code{\link{impute_time}()}
}
\concept{impute_time}
