% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/includes.R
\name{includes.partial_time.partial_time}
\alias{includes.partial_time.partial_time}
\title{Test for whether a timestamp could be included within parttime uncertainty}
\usage{
\method{includes.partial_time}{partial_time}(e1, e2)
}
\arguments{
\item{e1}{object to test whether includes \code{e2}}

\item{e2}{object to test whether included in \code{e1}}
}
\value{
A logical vector indicating whether \code{partial_time} \code{e1} includes
\code{partial_time} \code{e2}
}
\description{
Test for whether a timestamp could be included within parttime uncertainty
}
\examples{
x_chr <- c("2019", "2019-03-01", "2019-03",    "2018",    "",     "2018", "")
y_chr <- c("2019", "2019-03",    "2019-03-01", "2016-05", "2018", "",     "")

x <- as.parttime(x_chr)
y <- as.parttime(y_chr)

includes(x, y)

}
