% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/includes.R
\name{includes.partial_time}
\alias{includes.partial_time}
\title{Determine whether a partial time contains an object}
\usage{
\method{includes}{partial_time}(e1, e2)
}
\arguments{
\item{e1}{object to test whether includes \code{e2}}

\item{e2}{object to test whether included in \code{e1}}
}
\description{
Determine whether a partial time contains an object
}
