% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/includes.R
\name{includes}
\alias{includes}
\alias{includes.partial_time}
\title{Determine whether one object includes another}
\usage{
includes(e1, e2)

includes.partial_time(e1, e2)
}
\arguments{
\item{e1}{object to test whether includes \code{e2}}

\item{e2}{object to test whether included in \code{e1}}
}
\description{
Determine whether one object includes another

Determine whether a partial time contains an object
}
