% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_partial_time_is.R
\name{is.na.partial_time}
\alias{is.na.partial_time}
\title{Check if elements of a partial time vector is NA}
\usage{
\method{is.na}{partial_time}(x, ...)
}
\arguments{
\item{x}{partial_time vector to test}

\item{...}{additional arguments unused}
}
\value{
A logical vector indicating whether each element in the
\code{partial_time} vector is \code{NA}.
}
\description{
Check if elements of a partial time vector is NA
}
