% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_timespan_is.R
\name{is.na.timespan}
\alias{is.na.timespan}
\title{Check if elements of a partial time vector is NA}
\usage{
\method{is.na}{timespan}(x, ...)
}
\arguments{
\item{x}{partial_time vector to test}

\item{...}{additional arguments unused}
}
\description{
Check if elements of a partial time vector is NA
}
