% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_partial_time_is.R
\name{is.partial_time}
\alias{is.partial_time}
\title{Shorthand for checking partial time inheritance}
\usage{
is.partial_time(x)
}
\arguments{
\item{x}{object to test}
}
\description{
Shorthand for checking partial time inheritance

Test whether a partial_time object's time components are incomplete
}
