% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_timespan_is.R
\name{is.timespan}
\alias{is.timespan}
\title{Shorthand for checking timespan inheritance}
\usage{
is.timespan(x)
}
\arguments{
\item{x}{object to test}
}
\value{
A logical scalar indicating whether an object is a \code{partial_timespan}
object.
}
\description{
Shorthand for checking timespan inheritance
}
