% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_partial.R
\name{is_partial}
\alias{is_partial}
\title{Test whether a partial_time object is incomplete}
\usage{
is_partial(x, ..., components = c("year", "month", "day", "hour", "min",
  "sec", "secfrac", "tzhour", "tzmin"))
}
\arguments{
\item{x}{a partial_time object to test for incompleteness}

\item{...}{additional arguments unused}

\item{components}{components to include in testing}
}
\description{
Test whether a partial_time object is incomplete
}
