% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_partial.R
\name{is_partial_date}
\alias{is_partial_date}
\title{Test whether a partial_time object's date components are incomplete}
\usage{
is_partial_date(x, ...)
}
\arguments{
\item{x}{a partial_time object to test for incompleteness}

\item{...}{additional arguments unused}
}
\description{
Test whether a partial_time object's date components are incomplete
}
