% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_partial_time_is.R
\name{is_partial_time}
\alias{is_partial_time}
\title{Shorthand for checking partial time inheritance}
\usage{
is_partial_time(x)
}
\arguments{
\item{x}{object to test}
}
\description{
Shorthand for checking partial time inheritance
}
\concept{is_parttime}
