% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_timespan_is.R
\name{is_timespan}
\alias{is_timespan}
\title{Shorthand for checking timespan inheritance}
\usage{
is_timespan(x)
}
\arguments{
\item{x}{object to test}
}
\description{
Shorthand for checking timespan inheritance
}
