% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_iso8601_to_matrix.R
\docType{data}
\name{iso_regex}
\alias{iso_regex}
\title{slightly modified from parsedate - added 'secfrac' capture group}
\format{An object of class \code{character} of length 1.}
\usage{
iso_regex
}
\description{
slightly modified from parsedate - added 'secfrac' capture group
}
\keyword{datasets}
