% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_iso8601_to_matrix.R
\name{match_iso8601_to_matrix}
\alias{match_iso8601_to_matrix}
\title{A rework of parsedate::regexpr_to_df}
\usage{
match_iso8601_to_matrix(dates)
}
\arguments{
\item{dates}{character vector of dates to parse for iso8601 components}
}
\description{
Provides a vertical dataframe of matches more ammeneable to a tidyverse-style
analysis
}
