% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_partial_time_min_max.R
\name{max.partial_time}
\alias{max.partial_time}
\title{Get the maximum of a vector of partial_time objects}
\usage{
\method{max}{partial_time}(..., na.rm = FALSE, na.warn = TRUE)
}
\description{
Get the maximum of a vector of partial_time objects
}
\examples{
max(parttime(c("2019", "2018", "2019-02", "2018-03")))

}
