% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_partial_time_min_max.R
\name{min.partial_time}
\alias{min.partial_time}
\title{Get the minimum of a vector of partial_time objects}
\usage{
\method{min}{partial_time}(..., na.rm = FALSE, na.warn = TRUE)
}
\description{
Get the minimum of a vector of partial_time objects
}
