% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_month_day.R
\name{normalize_month_day}
\alias{normalize_month_day}
\title{Normalize days in month back to day limit for a given month}
\usage{
normalize_month_day(x)
}
\arguments{
\item{x}{a vector of parttime objects with days which may exceed viable days
of month}
}
\value{
a vector of parttime objects with normalized days of the month
}
\description{
Normalize days in month back to day limit for a given month
}
\examples{
x <- as.parttime(c("2019", "2019-02-31", "2019-01-05", "2016-02-31", "2016-01-05"))
parttime:::normalize_month_day(x)

}
