% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_partial_time_compat_vctrs.R
\name{obj_print_data.partial_time}
\alias{obj_print_data.partial_time}
\title{parttime data output}
\usage{
\method{obj_print_data}{partial_time}(x, ...)
}
\arguments{
\item{x}{A \code{partial_time} object}

\item{...}{Additional arguments unused}
}
\value{
A \code{character} representation of \code{partial_time}
}
\description{
parttime data output
}
