% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_partial_time_compat_vctrs.R
\name{obj_print_footer.partial_time}
\alias{obj_print_footer.partial_time}
\title{parttime footer}
\usage{
\method{obj_print_footer}{partial_time}(x, ...)
}
\arguments{
\item{x}{A \code{partial_time} object}

\item{...}{Additional arguments unused}
}
\value{
A string output when \code{partial_time} vector printing exceeds max print
length.
}
\description{
parttime footer
}
