% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_partial_time_compat_vctrs.R
\name{obj_print_header.partial_time}
\alias{obj_print_header.partial_time}
\title{parttime output header}
\usage{
\method{obj_print_header}{partial_time}(x, ...)
}
\arguments{
\item{x}{A \code{partial_time} object}

\item{...}{Additional arguments unused}
}
\value{
A \code{character} representation of \code{partial_time} metadata, as used to
describe its vector output header
}
\description{
parttime output header
}
