% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_to_parttime_matrix.R
\name{parse_failure_message}
\alias{parse_failure_message}
\title{Format a message communicating parse failure information}
\usage{
parse_failure_message(input, output)
}
\arguments{
\item{input}{The input provided to the parser}

\item{output}{The parttimes returned by the parser}
}
\value{
A message communicating failure modes for NAs introduced into output
that were not present in the input
}
\description{
Format a message communicating parse failure information
}
\keyword{internal}
