% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_iso8601.R
\name{parse_iso8601_matrix}
\alias{parse_iso8601_matrix}
\title{Parse an iso8601 datetime to a parttime-like matrix}
\usage{
parse_iso8601_matrix(dates)
}
\arguments{
\item{dates}{character vector of dates to parse for iso8601 components}
}
\description{
Parse an iso8601 datetime to a parttime-like matrix
}
\note{
In addition to parttime matrix fields, the returned matrix has
additional columns for alternative iso8601 formats such as \code{yearday},
\code{yearweek} and code{weakday}.
}
\keyword{internal}
