% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_iso8601.R, R/parse_to_parttime_matrix.R
\name{parse_iso8601}
\alias{parse_iso8601}
\alias{parse_to_parttime_matrix}
\title{Parse a character vector to create a matrix of datetime fields}
\usage{
parse_iso8601(dates, warn = TRUE)

parse_to_parttime_matrix(dates, regex = re_iso8601)
}
\arguments{
\item{dates}{character vector of dates to parse for iso8601 components}

\item{regex}{A regular expression used for parsing parttime strings. Defaults
to a parser for the ISO8601 standard. Should include named capture groups
for each datetime field. See Details for more information.}
}
\description{
Parse a character vector to create a matrix of datetime fields
}
\details{
A format regular expression should separate the components of a valid
datetime string, capturing strings which can be coerced to numeric values for
each of the nine datetime fields:

\itemize{
\item{\code{year}}
\item{\code{month}}
\item{\code{day}}
\item{\code{hour}}
\item{\code{min}}
\item{\code{sec}}
\item{\code{secfrac}}
\item{\code{tzhour}}
\item{\code{tzmin}}
}
}
\keyword{internal}
