% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_iso8601.R
\name{parse_iso8601_as_timespan}
\alias{parse_iso8601_as_timespan}
\title{Parse iso8601 datetime strings as timespan array}
\usage{
parse_iso8601_as_timespan(x, ...)
}
\arguments{
\item{x}{A \code{character} vector of iso8601 datetime strings}

\item{...}{Additional arguments unused}
}
\description{
Parse iso8601 datetime strings as timespan array
}
\note{
A timespan array is an internal data structure used as the backend
representation of timespan objects. It consists of two parttime-like matrices
(with the addition of an "inclusive" column), one for the lower- and
upper-bounds of the timespan. Collectively, this amounts to a three
dimensional array.
}
\keyword{internal}
