% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_partial_time_compat_lubridate.R
\name{parttime_access_and_assign}
\alias{parttime_access_and_assign}
\alias{year}
\alias{year<-}
\alias{month}
\alias{month<-}
\alias{mday}
\alias{mday<-}
\alias{day}
\alias{day<-}
\alias{hour}
\alias{hour<-}
\alias{minute}
\alias{minute<-}
\alias{second}
\alias{second<-}
\alias{tz}
\alias{tz<-}
\alias{year.partial_time}
\alias{year<-.partial_time}
\alias{year<-,partial_time-method}
\alias{month.partial_time}
\alias{month<-.partial_time}
\alias{month<-,partial_time-method}
\alias{mday.partial_time}
\alias{day<-.partial_time}
\alias{day<-,partial_time-method}
\alias{hour.partial_time}
\alias{hour<-.partial_time}
\alias{hour<-,partial_time-method}
\alias{minute.partial_time}
\alias{minute<-.partial_time}
\alias{minute<-,partial_time-method}
\alias{second.partial_time}
\alias{second<-.partial_time}
\alias{second<-,partial_time-method}
\alias{tz.partial_time}
\alias{tz<-.partial_time}
\title{Datetime component access and assignment functions}
\usage{
year(x)

year(x) <- value

month(x)

month(x) <- value

mday(x)

mday(x) <- value

day(x)

day(x) <- value

hour(x)

hour(x) <- value

minute(x)

minute(x) <- value

second(x)

second(x) <- value

tz(x)

tz(x) <- value

\method{year}{partial_time}(x)

\method{year}{partial_time}(x) <- value

\S4method{year}{partial_time}(x) <- value

\method{month}{partial_time}(x)

\method{month}{partial_time}(x) <- value

\S4method{month}{partial_time}(x) <- value

\method{mday}{partial_time}(x)

\method{day}{partial_time}(x) <- value

\S4method{day}{partial_time}(x) <- value

\method{hour}{partial_time}(x)

\method{hour}{partial_time}(x) <- value

\S4method{hour}{partial_time}(x) <- value

\method{minute}{partial_time}(x)

\method{minute}{partial_time}(x) <- value

\S4method{minute}{partial_time}(x) <- value

\method{second}{partial_time}(x)

\method{second}{partial_time}(x) <- value

\S4method{second}{partial_time}(x) <- value

\method{tz}{partial_time}(x)

\method{tz}{partial_time}(x) <- value
}
\arguments{
\item{x}{A time-like object to access or assign to}

\item{value}{For assignment, a value to assign}
}
\description{
Datetime component access and assignment functions
}
\note{
Care is taken to make these functionas as compatible as possible with similar
datetime packages. However, some functions may be masked and cause errors
using their masking functions.
}
