% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_partial_time_logical.R
\name{parttime_logical}
\alias{parttime_logical}
\title{This class is purely for retaining logical information for immediate
truncation as to not overccomplicate coersion needs with another vctr class}
\usage{
parttime_logical(x)
}
\description{
This class is purely for retaining logical information for immediate
truncation as to not overccomplicate coersion needs with another vctr class
}
