% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_partial_time_compat_pillar.R
\name{pillar_shaft.partial_time}
\alias{pillar_shaft.partial_time}
\title{parttime as pillar shaft}
\usage{
\method{pillar_shaft}{partial_time}(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Passed on to \code{\link[pillar:new_pillar_shaft]{new_pillar_shaft()}}.}
}
\value{
A \code{character} representation of a \code{partial_time} vector
}
\description{
parttime as pillar shaft
}
