% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_partial_time_compat_lubridate.R
\name{+,partial_time,Period-method}
\alias{+,partial_time,Period-method}
\title{Addition of a lubridate Period to a parttime partial_time}
\usage{
\S4method{+}{partial_time,Period}(e1, e2)
}
\arguments{
\item{e1}{A partial_time object}

\item{e2}{A lubridate Period object}
}
\description{
Addition of a lubridate Period to a parttime partial_time
}
