% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_partial_time_min_max.R
\name{pmax.partial_time}
\alias{pmax.partial_time}
\title{Get the elementwise maximum of parttime vectors}
\usage{
\method{pmax}{partial_time}(..., na.rm = FALSE)
}
\arguments{
\item{...}{numeric or character arguments (see Note).}

\item{na.rm}{a logical (\code{TRUE} or \code{FALSE}) indicating
    whether missing values should be removed.}
}
\value{
A \code{partial_time} vector with length equal to the maximum length of
the vectors provided where each value is the maximum of the recycled values
of each vector argument.
}
\description{
Get the elementwise maximum of parttime vectors
}
\examples{
pmax(
  parttime(c("2019", "2018", "2019-02", "2018",    "2010")),
  parttime(c("2020", NA,     "2019-03", "2018-01", "2010"))
)

}
