% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_partial_time_min_max.R
\name{pmin.partial_time}
\alias{pmin.partial_time}
\title{Get the elementwise minimum of parttime vectors}
\usage{
\method{pmin}{partial_time}(..., na.rm = FALSE)
}
\arguments{
\item{...}{numeric or character arguments (see Note).}

\item{na.rm}{a logical (\code{TRUE} or \code{FALSE}) indicating
    whether missing values should be removed.}
}
\value{
A \code{partial_time} vector with length equal to the maximum length of
the vectors provided where each value is the minimum of the recycled values
of each vector argument.
}
\description{
Get the elementwise minimum of parttime vectors
}
