% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propegate_na.R
\name{propegate_na}
\alias{propegate_na}
\title{Propegate field missingness from higher to lower resolution}
\usage{
propegate_na(x, keep_tz = FALSE)
}
\arguments{
\item{x}{a vector of parttime objects with days which may exceed viable days
of month}

\item{keep_tz}{logical indicating whether to retain timezone fields, even if
other fields are missing}
}
\value{
a vector of parttime objects with missingness propegated to lower
  resolution time fields
}
\description{
Propegate field missingness from higher to lower resolution
}
\examples{
x <- as.parttime(c("2019", "2019-02-31", "2019-01-05", "2016-02-31", 
  "2016-01-05", "2020-01-01 03:04:05.1234"))
vctrs::field(x, "pttm_mat")[,"min"] <- 23
parttime:::propegate_na(x)

}
