% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_cdisc.R
\name{re_cdisc_datetime}
\alias{re_cdisc_datetime}
\title{Regular expression for CDISC-style datetime parsing}
\usage{
re_cdisc_datetime
}
\description{
Regular expression for CDISC-style datetime parsing
}
\keyword{internal}
