% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_iso8601.R
\name{re_iso8601}
\alias{re_iso8601}
\title{slightly modified from parsedate - added 'secfrac' capture group}
\usage{
re_iso8601
}
\description{
slightly modified from parsedate - added 'secfrac' capture group
}
\keyword{internal}
