% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_to_parttime_matrix.R
\name{sample_partial_styles}
\alias{sample_partial_styles}
\title{Find unique forms of missingness}
\usage{
sample_partial_styles(x)
}
\arguments{
\item{x}{A \code{partial_time} vector}
}
\value{
The indices of the first instance of unique forms of missingness
}
\description{
Find unique forms of missingness
}
\examples{
x <- as.parttime(c("2022", "2023", "2021-01", "2024-12"))
x[sample_partial_styles(x)]

}
\keyword{internal}
