% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_timespan_compat_stats.R
\name{start}
\alias{start}
\title{start S3 generic}
\usage{
start(x, ...)
}
\arguments{
\item{x}{An object to retrieve the start from}

\item{...}{Additional arguments passed to methods}
}
\description{
A generic method to retrieve the start of an object
}
