% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_partial_time_extract.R
\name{[<-.partial_time}
\alias{[<-.partial_time}
\title{Subsetting assignment of partial_time objects}
\usage{
\method{[}{partial_time}(x, i, ...) <- value
}
\arguments{
\item{x}{an object from which to extract element(s) or in which to replace
element(s).}

\item{i}{indicies specifying elements to extract or replace. For further
details, see \link[base]{Extract}.}

\item{...}{arguments unused}

\item{value}{typically an array-like \R object of a similar class as
    \code{x}.}
}
\description{
Subsetting assignment of partial_time objects
}
\examples{
x <- as.parttime(c("2019", "2019-02", "2019-02-02"))
x[c(1, 3)] <- as.parttime(c("2000", "1999"))

}
