% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_timespan.R
\name{timespan}
\alias{timespan}
\title{Create a timespan object}
\usage{
timespan(start, end, inclusive = c(TRUE, FALSE))
}
\arguments{
\item{start}{vector of datetime objects to start timespans}

\item{end}{vector of datetime objects to end timespans}

\item{inclusive}{vector or matrix of logicals where each row is composed of
two logical values indicating whether the timespan start and end are
inclusive respectively}
}
\description{
Create a timespan object
}
