% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_gmt.R
\name{to_gmt}
\alias{to_gmt}
\title{Generic for coercing timestamps to GMT timezone}
\usage{
to_gmt(x)
}
\arguments{
\item{x}{object to coerce to GMT time}
}
\value{
A time object adjusted to GMT time
}
\description{
Generic for coercing timestamps to GMT timezone
}
