% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{Shorten a timespan}
\usage{
trim(x, ...)
}
\arguments{
\item{x}{timespan object to trim}

\item{...}{additional arguments passed on to functions}
}
\description{
Shorten a timespan
}
