% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_partial_time_compat_pillar.R
\name{type_sum.partial_time}
\alias{type_sum.partial_time}
\title{parttime type name}
\usage{
\method{type_sum}{partial_time}(x)
}
\arguments{
\item{x}{an object to summarise. Generally only methods of atomic vectors
and variants have been implemented.}
}
\value{
A \code{character} scalar shorthand representation of the \code{partial_time}
class name
}
\description{
parttime type name
}
