% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_partial_time_coersion.R
\name{vec_cast.partial_time.matrix}
\alias{vec_cast.partial_time.matrix}
\title{Cast a matrix to a partial time}
\usage{
\method{vec_cast.partial_time}{matrix}(x, to, ...)
}
\arguments{
\item{x}{Vectors to cast.}

\item{to}{Type to cast to. If \code{NULL}, \code{x} will be returned as is.}

\item{...}{For \code{vec_cast_common()}, vectors to cast. For
\code{vec_cast()} and \code{vec_restore()}, these dots are only for future
extensions and should be empty.}
}
\description{
Cast a matrix to a partial time
}
