% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_timespan_coercion.R
\name{vec_cast.timespan.numeric}
\alias{vec_cast.timespan.numeric}
\title{Cast an array to a timespan}
\usage{
\method{vec_cast}{timespan.numeric}(x, to, ...)
}
\arguments{
\item{x}{Vectors to cast.}

\item{to}{Type to cast to. If \code{NULL}, \code{x} will be returned as is.}

\item{...}{For \code{vec_cast_common()}, vectors to cast. For
\code{vec_cast()}, \code{vec_cast_default()}, and \code{vec_restore()}, these
dots are only for future extensions and should be empty.}
}
\value{
A \code{partial_timespan} vector
}
\description{
Cast an array to a timespan
}
