% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_timespan_coersion.R
\name{vec_cast.timespan.partial_time}
\alias{vec_cast.timespan.partial_time}
\title{Cast partial time to timespan, representing uncertainty as a range}
\usage{
\method{vec_cast}{timespan.partial_time}(x, to, ...)
}
\arguments{
\item{x}{Vectors to cast.}

\item{to}{Type to cast to. If \code{NULL}, \code{x} will be returned as is.}

\item{...}{For \code{vec_cast_common()}, vectors to cast. For
\code{vec_cast()}, \code{vec_cast_default()}, and \code{vec_restore()}, these
dots are only for future extensions and should be empty.}
}
\description{
Cast partial time to timespan, representing uncertainty as a range
}
