% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_partial_time_compat_vctrs.R
\name{vec_ptype_full.partial_time}
\alias{vec_ptype_full.partial_time}
\title{Full parttime class name}
\usage{
\method{vec_ptype_full}{partial_time}(x)
}
\arguments{
\item{x}{A \code{partial_time} object}
}
\value{
A \code{character} representation of the \code{partial_time} class name
}
\description{
Full parttime class name
}
