% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_partial.R
\name{has_partial_date}
\alias{has_partial_date}
\title{Test whether a partial_time object's date components are incomplete}
\usage{
has_partial_date(x)
}
\arguments{
\item{x}{a partial_time object to test for incompleteness}
}
\value{
A logical vector indicating whether each element of a \code{partial_time}
has any missing date fields.
}
\description{
Test whether a partial_time object's date components are incomplete
}
