% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_time.R
\name{impute_time}
\alias{impute_time}
\title{Impute a partial time object with a timestamp or specific fields}
\usage{
impute_time(x, time, tz, ...)
}
\arguments{
\item{x}{a datetime-like object to impute}

\item{time}{a datetime-like object to use for imputation}

\item{tz}{a character timezone name for imputation, a character value to use
as the timezone part of the datetime or an numeric minute offset.}

\item{...}{additional individual named fields to impute. Can be one of
"year", "month", "day", "hour", "min", "sec", "secfrac", "tzhour", "tzmin".}
}
\value{
a new partial_time with specified fields imputed
}
\description{
Impute a partial time object with a timestamp or specific fields
}
