% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/includes.R
\name{includes.partial_time.partial_time}
\alias{includes.partial_time.partial_time}
\title{Test for whether a timestamp could be included within parttime uncertainty}
\usage{
\method{includes}{partial_time.partial_time}(e1, e2)
}
\arguments{
\item{e1}{object to test whether includes \code{e2}}

\item{e2}{object to test whether included in \code{e1}}
}
\description{
Test for whether a timestamp could be included within parttime uncertainty
}
\examples{
x <- as.parttime(c("2019", "2019-03-01", "2019-03",    "2018",    "",     "2018", ""))
y <- as.parttime(c("2019", "2019-03",    "2019-03-01", "2016-05", "2018", "",     ""))

includes(x, y)

}
