% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/includes.R
\name{includes}
\alias{includes}
\title{Determine whether one object includes another}
\usage{
includes(e1, e2)
}
\arguments{
\item{e1}{object to test whether includes \code{e2}}

\item{e2}{object to test whether included in \code{e1}}
}
\value{
A logical vector indicating whether \code{e1} includes \code{e2}
}
\description{
Determine whether one object includes another
}
