% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_partial_time_is.R, R/has_partial.R
\name{is_partial_time}
\alias{is_partial_time}
\title{Shorthand for checking partial time inheritance}
\usage{
is_partial_time(x, ...)

is_partial_time(x, ...)
}
\arguments{
\item{x}{object to test}

\item{...}{additional arguments unused}
}
\description{
Shorthand for checking partial time inheritance

Test whether a partial_time object's time components are incomplete
}
\concept{is_parttime}
