% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_partial_time_is.R
\name{is_partial_time}
\alias{is_partial_time}
\alias{is.partial_time}
\alias{is_parttime}
\alias{is.parttime}
\title{Shorthand for checking partial time inheritance}
\usage{
is_partial_time(x)

is.partial_time(x)

is_parttime(x)

is.parttime(x)
}
\arguments{
\item{x}{object to test}
}
\value{
A logical scalar indicating whether an object is a \code{partial_time}
object.
}
\description{
Shorthand for checking partial time inheritance
}
\concept{is_parttime}
