% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_partial_time_min_max.R
\name{min.partial_time}
\alias{min.partial_time}
\title{Get the minimum of aparttime vector}
\usage{
\method{min}{partial_time}(..., na.rm = FALSE, na.warn = TRUE)
}
\arguments{
\item{...}{partial_time objects}

\item{na.rm}{whether \code{NA} should be removed when calculating min}

\item{na.warn}{whether to raise a warning for \code{NA}}
}
\value{
A \code{partial_time} scalar
}
\description{
Get the minimum of aparttime vector
}
