% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_partial_time_ops.R
\name{neq_parttimes}
\alias{neq_parttimes}
\title{Not-equal comparison handler for partial_time objects}
\usage{
neq_parttimes(generic, e1, e2)
}
\description{
Not-equal comparison handler for partial_time objects
}
\examples{
x <- as.parttime(c("2019", "2018-01",    NA, "2011"))
y <- as.parttime(c("2019", "2018-01-03", NA, "2010-01"))

x != y

}
