% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_partial_time_ops.R
\name{Ops.timespan}
\alias{Ops.timespan}
\title{Handler for Ops generics for timespan objects}
\usage{
\method{Ops}{timespan}(e1, e2)
}
\arguments{
\item{e1}{objects}

\item{e2}{objects}
}
\description{
Handler for Ops generics for timespan objects
}
