% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_partial_time.R
\name{parttime}
\alias{parttime}
\title{Create a parttime object}
\usage{
parttime(year = NA, month = NA, day = NA, hour = NA, min = NA,
  sec = NA,
  tzhour = interpret_tz(getOption("parttime.assume_tz_offset",
  NA))\%/\%60,
  tzmin = interpret_tz(getOption("parttime.assume_tz_offset", NA))\%\%60)
}
\description{
Create a parttime object
}
\examples{
parttime(2019)

}
