% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_partial_time_min_max.R
\name{pmax.partial_time}
\alias{pmax.partial_time}
\title{Get the elementwise maximum of vectors of partial_time objects}
\usage{
\method{pmax}{partial_time}(..., na.rm = FALSE)
}
\arguments{
\item{...}{numeric or character arguments (see Note).}

\item{na.rm}{a logical indicating whether missing values should be
    removed.}
}
\description{
Get the elementwise maximum of vectors of partial_time objects
}
\examples{
pmax(
  parttime(c("2019", "2018", "2019-02", "2018",    "2010")),
  parttime(c("2020", NA,     "2019-03", "2018-01", "2010"))
)

}
