% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_partial_time_min_max.R
\name{pmin.partial_time}
\alias{pmin.partial_time}
\title{Get the elementwise minimum of vectors of partial_time objects}
\usage{
\method{pmin}{partial_time}(..., na.rm = FALSE)
}
\description{
Get the elementwise minimum of vectors of partial_time objects
}
