% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_cdisc.R
\docType{data}
\name{re_cdisc_datetime}
\alias{re_cdisc_datetime}
\title{Regular expression for CDISC-style datetime parsing}
\format{
An object of class \code{character} of length 1.
}
\usage{
re_cdisc_datetime
}
\description{
Regular expression for CDISC-style datetime parsing
}
\keyword{internal}
