% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_iso8601.R, R/parse_iso8601.R
\docType{data}
\name{re_iso8601}
\alias{re_iso8601}
\title{slightly modified from parsedate - added 'secfrac' capture group}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
re_iso8601

re_iso8601
}
\description{
slightly modified from parsedate - added 'secfrac' capture group

slightly modified from parsedate - added 'secfrac' capture group
}
\keyword{datasets}
