% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_partial_time_compat_vctrs.R
\name{vec_ptype_abbr.partial_time}
\alias{vec_ptype_abbr.partial_time}
\title{Abbreviated partial time class name}
\usage{
\method{vec_ptype_abbr}{partial_time}(x)
}
\arguments{
\item{x}{A \code{partial_time} object}
}
\value{
A \code{character} representation of the abbreviated \code{partial_time} class
name
}
\description{
Abbreviated partial time class name
}
